requirements.txt
pycryptodome==3.20.0
Flask==3.0.3
requests==2.32.3
flask-cors==5.0.0
flask-swagger-ui==4.11.1